# Proposal(hi): Make Hi faster (at generating teeth)