--- a/lerna.json
+++ b/lerna.json
@@ -1,5 +1,5 @@
 {
-  "version": "2.21.0-alpha.0",
+  "version": "2.21.0-rc.0",
   "npmClient": "yarn",
   "command": {
     "publish": {