--- a/packages/freesewing.shared/config/designs.json
+++ b/packages/freesewing.shared/config/designs.json
@@ -2,6 +2,7 @@
   "accessories": [
     "florence",
     "florent",
+    "hi",
     "holmes",
     "hortensia",
     "lucy",
@@ -41,6 +42,7 @@
     "teagan",
     "theo",
     "tiberius",
+    "unice",
     "ursula",
     "wahid",
     "walburga",