@woutervdub 2022-06-15
The shape of the curve is insignificant IMHO. Mirroring, and therefore requiring an even amount of teeth is also not an issue. And allows us to say that it is not a complete copy. 

Why did the size of the small teeth (extreme sides of the mouth) change?