@joostdecock 2022-06-16
@woutervdub I'm going to merge this now because I have another big set of changes and letting things diverge will be a nightmare later.

However, we can continue to fine-tune this prior to publishing it. The variation of tooth size for example is something we can review.